refactor : AuthorBox 컴포넌트
