+++ frontend/components/banner/BannerSlider.tsx
