+++ frontend/components/banner/BannerText.tsx
