+++ frontend/components/common_box/GridBox.tsx
