+++ frontend/pages/project/[id]/index.tsx
@@ -1,15 +1,12 @@
-import GridBox from '@/components/GridBox';
 import Tag from '@/components/Tag';
 import PeriodBox from '@/components/project/PeriodBox';
 import TagBox from '@/components/project/TagBox';
 import dynamic from 'next/dynamic';
 import styled from 'styled-components';
 import ContentSkeleton from '@/components/skeleton/ContentSkeleton';
-import { AiFillHeart, AiOutlineHeart } from 'react-icons/ai';
 import { formatDate2 } from '@/util/date';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import Position from '@/components/Position';
 import Message from '@/components/Message';
 import { useRecoilValue } from 'recoil';
 import { loggedInUserState } from '@/recoil/atom';
@@ -17,7 +14,9 @@ import { BUTTON_STATE } from '@/constant/constant';
 import { useProject } from '@/hooks/react-query/project/useProject';
 import AnswerBox from '@/components/answer/AnswerBox';
 import StacksBox from '@/components/project/StacksBox';
-import HeartBox from '@/components/HeartBox';
+import HeartBox from '@/components/common_box/HeartBox';
+import GridBox from '@/components/common_box/GridBox';
+import AuthorBox from '@/components/common_box/AuthorBox';
 const ReactMarkdown = dynamic(() => import('@/components/editor/ContentBox'), {
   ssr: false,
   loading: () => <ContentSkeleton />,
@@ -33,9 +32,6 @@ const ViewProject = () => {
       behavior: 'smooth',
     });
   }, [router]);
-
-  //임시 데이터
-  const [userHeart, setUserHeart] = useState(false);
 
   //프로젝트 데이터 요청
   const {
@@ -84,53 +80,12 @@ const ViewProject = () => {
       ) : (
         <>
           <Side>
-            <div className="author-box">
-              <div className="author noto-medium">
-                <div className="top">
-                  <img
-                    src="https://noticon-static.tammolo.com/dgggcrkxq/image/upload/v1567008394/noticon/ohybolu4ensol1gzqas1.png"
-                    alt="author"
-                  />
-                  <div className="user-title">{data.memberInfo.name}</div>
-                  <div className="noto-medium">
-                    <Position text={data.writerPosition} />
-                  </div>
-                  {data.memberInfo.email !== loggedInUser?.email && (
-                    <div
-                      className="saveStar"
-                      onClick={() => setUserHeart(!userHeart)}
-                    >
-                      <span className="icon-box">
-                        {userHeart ? (
-                          <AiOutlineHeart fill={'#ececec'} />
-                        ) : (
-                          <AiFillHeart fill="red" />
-                        )}
-                      </span>
-                    </div>
-                  )}
-                </div>
-                <div className="detail-box">
-                  <div className="detail-sub-box">
-                    <div className="detail-num">
-                      {/* 서버 데이터가 들어오면 작업해줘야할 부분 */}
-                      {3} <span>개</span>
-                    </div>
-                    <div className="detail-title">진행 프로젝트</div>
-                  </div>
-                  <div className="center-border"></div>
-                  <div className="detail-sub-box">
-                    <div className="detail-num">
-                      {data.memberInfo.totalStar} <span>개</span>
-                    </div>
-                    <div className="detail-title">평가 점수</div>
-                  </div>
-                </div>
-                {data.memberInfo.email !== loggedInUser?.email && (
-                  <Tag>쪽지 보내기</Tag>
-                )}
-              </div>
-            </div>
+            <AuthorBox
+              userImg={data.memberInfo.profileImageUrl}
+              userName={data.memberInfo.name}
+              isAuthor={data.memberInfo.email !== loggedInUser?.email}
+              totalStar={data.memberInfo.totalStar}
+            />
             <PeriodBox
               start={new Date(data.startDate)}
               end={new Date(data.endDate)}
@@ -302,111 +257,6 @@ const Side = styled.div`
   align-items: center;
   gap: 32px;
   padding: var(--padding-1);
-
-  .author-box {
-    border: solid 2px #ececec;
-    padding: 40px 30px 20px;
-    border-radius: 15px;
-    display: flex;
-
-    .author {
-      display: flex;
-      align-items: center;
-      flex-direction: column;
-
-      .top {
-        position: relative;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        gap: 20px;
-        border-bottom: solid 2px #ececec;
-        padding-bottom: 20px;
-
-        > img {
-          border-radius: 50%;
-          box-shadow: 0px 0px 11px 11px rgba(234, 234, 234, 0.77);
-        }
-
-        > .user-title {
-          color: #9f9f9f;
-          font-size: 18px;
-        }
-
-        > .saveStar {
-          position: absolute;
-          width: 50px;
-          height: 50px;
-          background-color: #cecece;
-          border-radius: 50%;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          top: 0;
-          right: 0;
-          cursor: pointer;
-
-          > .icon-box {
-            width: 50%;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            font-size: 30px;
-          }
-        }
-      }
-
-      > .detail-box {
-        display: flex;
-        width: 80%;
-        justify-content: space-between;
-        margin-bottom: 24px;
-
-        > .center-border {
-          width: 1px;
-          height: 150%;
-          border: solid 1px #ececec;
-        }
-
-        > .detail-sub-box {
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: center;
-
-          > .detail-num {
-            padding: 10px;
-            font-size: 24px;
-
-            > span {
-              font-size: 15px;
-              color: #828282;
-            }
-          }
-
-          > .detail-title {
-            font-size: 12px;
-          }
-        }
-      }
-
-      > div:last-child {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 16px;
-        margin-top: 30px;
-
-        > div {
-          font-weight: 900;
-        }
-        .tag {
-          cursor: pointer;
-        }
-      }
-    }
-  }
 
   .want-box {
     width: 100%;
